+++ apps/docs/src/app/core/component-docs/dialog/dialog-docs.component.ts
@@ -34,6 +34,8 @@ import * as objectDialogHtml from '!raw-loader!./examples/dialog-object-example/
 import * as autoLabelTs from '!raw-loader!./examples/auto-label/auto-label-dialog-example.component.ts';
 import * as autoLabelHtml from '!raw-loader!./examples/auto-label/auto-label-dialog-example.component.html';
 
+import * as popoverDialogTs from '!raw-loader!./examples/dialog-inner-popover/dialog-inner-popover.component.ts';
+
 import { DialogService } from '@fundamental-ngx/core/dialog';
 import { ExampleFile } from '../../../documentation/core-helpers/code-example/example-file';
 import { SchemaFactoryService } from '../../../schema/services/schema-factory/schema-factory.service';
@@ -298,6 +300,15 @@ export class DialogDocsComponent {
             code: complexDialogTs,
             fileName: 'dialog-complex-example',
             component: 'DialogComplexExampleComponent'
+        }
+    ];
+
+    dialogInnerPopover: ExampleFile[] = [
+        {
+            language: 'typescript',
+            code: popoverDialogTs,
+            fileName: 'dialog-inner-popover',
+            component: 'DialogInnerPopoverComponent'
         }
     ];