+++ apps/docs/src/app/core/component-docs/moment-datetime-adapter/moment-datetime-adapter-header/moment-datetime-adapter-header.component.ts
@@ -1,0 +1,7 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'app-moment-datetime-adapter-header',
+    templateUrl: './moment-datetime-adapter-header.component.html'
+})
+export class MomentDatetimeAdapterHeaderComponent {}