+++ libs/platform/src/lib/components/table/table.component.spec.ts
@@ -196,7 +196,6 @@ describe('TableComponent internal', () => {
 
         expect(serviceFreezeToSpy).toHaveBeenCalledWith(columnKey);
         expect(component.freezeColumnsTo).toEqual(columnKey);
-        expect(component._selectionColumnWidth).toEqual(0);
         expect(component._tablePadding).toEqual(0);
     });
 });